Static Lib

+++ IBX_Engine/src/IBX/ImGui/ImGuiLayer.cpp
@@ -1,16 +1,16 @@
-#include "ibxpch.h";
+#include "ibxpch.h"
 
-#include "ImGuiLayer.h";
+#include "ImGuiLayer.h"
 
-#include "imgui.h";
-#include "backends/imgui_impl_glfw.h";
-#include "backends/imgui_impl_opengl3.h";
+#include "imgui.h"
+#include "backends/imgui_impl_glfw.h"
+#include "backends/imgui_impl_opengl3.h"
 
-#include "IBX/Application.h";
+#include "IBX/Application.h"
 
 // Temporary - should be removed
-#include <GLFW/glfw3.h>;
-#include <glad/glad.h>;
+#include <GLFW/glfw3.h>
+#include <glad/glad.h>
 
 namespace IBX_Engine
 {
@@ -76,7 +76,7 @@ namespace IBX_Engine
 	{
 		ImGuiIO& io = ImGui::GetIO();
 		Application& app = Application::Get();
-		io.DisplaySize = ImVec2(app.GetWindow().GetWidth(), app.GetWindow().GetHeight());
+		io.DisplaySize = ImVec2((float)app.GetWindow().GetWidth(), (float)app.GetWindow().GetHeight());
 
 		// Rendering
 		ImGui::Render();
